code :)
This is the code for my project, creates a file and puts everything on there. I'm sort of a beginner so any help would be much appreciated. Also, I did NOT write the formula (ln 1-13, ln 30-31), I am not sure who made the code (will put it in the readme if found). No clue how the formula works but it does and thats all that matters. Thank you!